+++ python/py_models.cpp
@@ -1,4 +1,4 @@
-/// Copyright (c) 2018-2020, Parker Owan.  All rights reserved.
+/// Copyright (c) 2018-2021, Parker Owan.  All rights reserved.
 /// Licensed under BSD-3 Clause, https://opensource.org/licenses/BSD-3-Clause
 
 #include "python/py_models.h"
@@ -12,6 +12,40 @@ void export_py_models(py::module& m_sup) {
            py::arg("control"), py::return_value_policy::reference_internal)
       .def("measurement", &sia::MarkovProcess::measurement, py::arg("state"),
            py::return_value_policy::reference_internal);
+
+  py::class_<sia::Trajectory>(m, "Trajectory")
+      .def_readwrite("states", &sia::Trajectory::states)
+      .def_readwrite("controls", &sia::Trajectory::controls)
+      .def_readwrite("measurements", &sia::Trajectory::measurements);
+
+  py::class_<sia::Trajectories>(m, "Trajectories")
+      .def(py::init<const std::vector<sia::Trajectory>&>(), py::arg("data"))
+      .def("data", &sia::Trajectories::data)
+      .def("size", &sia::Trajectories::size)
+      .def("states", &sia::Trajectories::states, py::arg("k"))
+      .def("controls", &sia::Trajectories::controls, py::arg("k"))
+      .def("measurements", &sia::Trajectories::measurements, py::arg("k"));
+
+  m.def("simulate",
+        static_cast<sia::Trajectory (*)(
+            sia::MarkovProcess&, const Eigen::VectorXd&, const Eigen::MatrixXd&,
+            bool)>(&sia::simulate),
+        py::arg("system"), py::arg("state"), py::arg("controls"),
+        py::arg("sample") = true);
+
+  m.def("simulate",
+        static_cast<sia::Trajectories (*)(
+            sia::MarkovProcess&, const std::vector<Eigen::VectorXd>&,
+            const Eigen::MatrixXd&, bool)>(&sia::simulate),
+        py::arg("system"), py::arg("states"), py::arg("controls"),
+        py::arg("sample") = true);
+
+  m.def("simulate",
+        static_cast<sia::Trajectories (*)(
+            sia::MarkovProcess&, const std::vector<Eigen::VectorXd>&,
+            const std::vector<Eigen::MatrixXd>&, bool)>(&sia::simulate),
+        py::arg("system"), py::arg("states"), py::arg("controls"),
+        py::arg("sample") = true);
 
   py::class_<sia::NonlinearGaussian, sia::MarkovProcess>(m, "NonlinearGaussian")
       .def(py::init<sia::DynamicsEquation, sia::MeasurementEquation,