+++ sia/belief/particles.h
@@ -1,4 +1,4 @@
-/// Copyright (c) 2018-2020, Parker Owan.  All rights reserved.
+/// Copyright (c) 2018-2021, Parker Owan.  All rights reserved.
 /// Licensed under BSD-3 Clause, https://opensource.org/licenses/BSD-3-Clause
 
 #pragma once
@@ -68,6 +68,9 @@ class Particles : public Distribution {
   /// Returns the unbiased sample covariance (or weighted covariance if
   /// WeightedStats is set).
   const Eigen::MatrixXd covariance() const override;
+
+  const Eigen::VectorXd vectorize() const override;
+  bool devectorize(const Eigen::VectorXd& data) override;
 
   bool getUseWeightedStats() const;
   void setUseWeightedStats(bool weighted_stats);